# Optimize CompositeBeanHelper performance with comprehensive caching