@gnodet 2025-07-17
## 💚 All backports created successfully

| Status | Branch | Result |
|:------:|:------:|:------|
|✅|maven-4.0.x|[<img src="https://img.shields.io/github/pulls/detail/state/apache/maven/10927">](https://github.com/apache/maven/pull/10927)|

### Questions ?
Please refer to the [Backport tool documentation](https://github.com/sorenlouv/backport)

<!--- Backport version: 10.0.1 -->