@slawekjaranowski 2025-07-17
@gnodet there is no category in release-drafter for performance label ....
https://github.com/apache/maven-gh-actions-shared/blob/main/.github/release-drafter.yml


should we add such?